Added some more notes to the README and renamed some modules
