+++ .travis.yml
@@ -11,7 +11,7 @@ notifications:
 
 env:
   global:
-    - HE_VERSION=0.14.8
+    - HE_VERSION=0.14.9
 
 branches:
   only:
