+++ lib/modules/node-chs/index.js
