+++ lib/modules/node-nsp/index.js
