+++ lib/modules/ruby-bundler-scan/index.js
